Structuring headings for SEO
On the home page (and list pages, for now) the blog title is the H1
heading, marking the title of the page. That's also where we want links to
articles, being their titles, but these titles are not he main H1 in the
page.

On the single page, however, the post's title becomes the page title and
doesn't need to be a link anymore.